third_party/blink/web_tests/fast/css-grid-layout/grid-auto-repeat-huge-grid-007.html

<!DOCTYPE html>
<title>Test for auto-fit and auto-fill with huge grids (lots of tracks)</title>
<link href="resources/grid.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/width-keyword-classes.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/height-keyword-classes.css" rel="stylesheet">
<link href="resources/huge-grids.css" rel="stylesheet">

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/grid-definitions-parsing-utils.js"></script>
<script src="resources/grid-tracks-length.js"></script>

<div id="handMadeAutoFillRows" class="grid tallGrid handMadeMoreThanThousandAutoFillRows"></div>
<div id="handMadeAutoFitRows" class="grid tallGrid handMadeMoreThanThousandAutoFitRows">
    <div>Item1</div>
    <div>Item2</div>
    <div class="lastRow">Item3</div>
</div>

<div id="handMadeAutoFillCols" class="grid wideGrid handMadeMoreThanThousandAutoFillCols"></div>
<div id="handMadeAutoFitCols" class="grid wideGrid handMadeMoreThanThousandAutoFitCols">
    <div>Item1</div>
    <div>Item2</div>
    <div class="lastColumn">Item3</div>
</div>

<script>

test(function() {
     var autoFillGrid = testElement("handMadeAutoFillRows", "grid-template-rows", 100000);
     var autoFitGrid = testElement("handMadeAutoFitRows", "grid-template-rows", 100000);

     assert_equals(autoFillGrid[100000 - 1], "10px");
     assert_equals(autoFillGrid[0], "10px");
     assert_equals(autoFitGrid[100000 - 1], "10px");
     assert_equals(autoFitGrid[0], "10px");
}, "Test that we don't crash when there are more than kGridMaxTracks rows in the auto repeat <track-list>.");

test(function() {
     var autoFillGrid = testElement("handMadeAutoFillCols", "grid-template-columns", 100000);
     var autoFitGrid = testElement("handMadeAutoFitCols", "grid-template-columns", 100000);

     assert_equals(autoFillGrid[100000 - 1], "10px");
     assert_equals(autoFillGrid[0], "10px");
     assert_equals(autoFitGrid[100000 - 1], "10px");
     assert_equals(autoFitGrid[0], "10px");
}, "Test that we don't crash when there are more than kGridMaxTracks columns in the auto repeat <track-list>.");

</script>
